@@ -6,8 +6,11 @@
   "private": true,
   "main": "public/electron.js",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "11.3.0",
+    "@mui/icons-material": "^5.0.0",
+    "@mui/material": "^5.0.0",
+    "@mui/styles": "^5.0.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -18,15 +21,15 @@
     "electron-is-dev": "^1.2.0",
     "lodash": "^4.17.21",
     "loglevel": "^1.7.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-dropzone": "^10.2.2",
     "react-hotkeys": "^2.0.0",
     "react-hotkeys-hook": "^1.6.1",
     "react-redux": "^7.2.5",
     "react-router-cache-route": "^1.11.1",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "3.4.0",
+    "react-scripts": "4.0.3",
     "react-selectable-fast": "^3.4.0",
     "react-sketch": "^0.5.1",
     "recharts": "^2.1.2",
