@@ -1,7 +1,7 @@
-import { createTheme } from '@material-ui/core/styles';
-import { blue, indigo } from '@material-ui/core/colors';
+import { createTheme, adaptV4Theme } from '@mui/material/styles';
+import { blue, indigo } from '@mui/material/colors';
 
-export const theme = createTheme({
+export const theme = createTheme(adaptV4Theme({
     palette: {
         primary: { light: "rgb(152, 177, 217)", main: "rgb(93, 115, 150)", dark: "rgb(31, 48, 68)" },
         secondary: { light: indigo[50], main: indigo[500], dark: indigo[700] },
@@ -15,7 +15,7 @@ export const theme = createTheme({
             'sans-serif'
         ].join(','),
     }
-});
+}));
 
 export const RECEIVER_CHANNEL = "fromMain";
 export const SENDER_CHANNEL = "toMain";
