@@ -4,8 +4,8 @@ import {
     AccordionDetails,
     AccordionSummary,
     Typography
-} from '@material-ui/core/';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+} from '@mui/material/';
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 
 // https://stackoverflow.com/a/34842797
 export const imageHash = s => '' + (s.split('').reduce((a, b) => (((a << 5) - a) + b.charCodeAt(0)) | 0, 0) >>> 0); //bit shift in order to get unsigned integer
