@@ -1,13 +1,14 @@
 import React, { useState } from 'react';
-import PlayPauseButton from './PlayPauseButton';
-import RestartButton from './RestartButton';
-import RewindButton from './RewindButton';
-import ForwardButton from './ForwardButton';
-import StopButton from './StopButton';
+import PlayPauseButton from './buttons/PlayPauseButton';
+import RestartButton from './buttons/RestartButton';
+import RewindButton from './buttons/RewindButton';
+import ForwardButton from './buttons/ForwardButton';
+import StopButton from './buttons/StopButton';
+import FigureDrawingSettings from './FigureDrawingSettings';
 
-const FigureDrawing = () => {
+const ActiveDrawingSessionLayout = () => {
     return (
-        <div className="figure-drawing-container">
+        <>
             <div className="current-image-container"></div>
             <div className="figure-drawing-buttons-container">
                 <RestartButton />
@@ -16,6 +17,16 @@ const FigureDrawing = () => {
                 <ForwardButton />
                 <StopButton />
             </div>
+        </>
+    );
+}
+
+const FigureDrawing = () => {
+    const [isSessionActive, setIsSessionActive] = useState(false);
+    const InnerComponent = isSessionActive ? ActiveDrawingSessionLayout : FigureDrawingSettings;
+    return (
+        <div className="figure-drawing-container">
+            <InnerComponent />
         </div>
     );
 }
